Add checking linkedBy for Aspects Add update free/non-free tests

+++ backend/src/main/kotlin/com/infowings/catalog/data/aspect/AspectService.kt
@@ -138,18 +138,10 @@ class AspectService(
 
             val refBook = referenceBookService.getReferenceBookOrNull(aspectId)
 
-            //TODO: checking if children items linked by Objects and set correct linkedRefBookItems!
-            val linkedRefBookItems: List<ReferenceBookItem> = emptyList()
-            val hasLinkedRefBookItem = linkedRefBookItems.isNotEmpty()
+            val linked = aspectVertex.isLinkedBy()
 
-            val linked = aspectVertex.isLinkedBy() || hasLinkedRefBookItem
             when {
-                aspectVertex.isLinkedBy() && force -> {
-                    historyService.storeFact(aspectVertex.toSoftDeleteFact(context))
-                    aspectDaoService.fakeRemove(aspectVertex)
-                }
                 linked && force -> {
-                    if (refBook != null) referenceBookService.removeReferenceBook(refBook, context.userVertex, force)
                     historyService.storeFact(aspectVertex.toSoftDeleteFact(context))
                     aspectDaoService.fakeRemove(aspectVertex)
                 }
@@ -228,6 +220,10 @@ class AspectService(
 
         val vertex = aspectDaoService.getAspectPropertyVertex(property.id)
                 ?: throw AspectPropertyDoesNotExist(property.id)
+
+        if (vertex.isLinkedBy()) {
+            throw AspectPropertyHasValueException(vertex.id)
+        }
 
         return@transaction aspectDaoService.remove(vertex)
     }
@@ -363,7 +359,7 @@ class AspectCyclicDependencyException(cyclicIds: List<String>) :
 
 class AspectNameCannotBeNull : AspectException()
 class AspectHasLinkedEntitiesException(val id: String) : AspectException("Some entities refer to aspect $id")
-
+class AspectPropertyHasValueException(val id: String) : AspectException("Aspect Property $id has a value")
 class AspectInconsistentStateException(message: String) : AspectException(message)
 
 class AspectEmptyChangeException : AspectException()
